feat: make methods consistent
